rearrange some stuff
